Refactor Input component and remove unused code
This commit refactors the Input component in the shared folder. It removes the unused Stack and Typography imports and cleans up the code. This change improves the codebase by removing unnecessary dependencies and reducing complexity.

Refactor ModalAddTime component and disable timeRanges input

This commit refactors the ModalAddTime component in the sn-budgeting folder. It removes the unused timeRanges input and disables it. This change improves the component by removing unnecessary code and improving the user experience.