Refactor API client and add file upload functionality
This commit refactors the API client in the `client.ts` file to import the `AxiosInstance` and `AxiosRequestConfig` from the `axios` library. It also adds a new method `uploadFile` to the `RequestClient` class, which allows uploading a file to the server.

In the `endpoint.ts` file, a new endpoint `UPLOAD_FILE` is added for file uploads.

In the `EmployeeDetailHeader.tsx` file, the code is modified to use the `uploadFile` method instead of the deprecated `upload` method to upload the avatar file. The `avatarUrl` is now obtained from the `data.link` property of the response.

These changes improve the code structure and introduce a more efficient way to handle file uploads.